+++ x-pack/solutions/observability/plugins/slo/public/components/slo/slo_badges/index.tsx
@@ -8,3 +8,4 @@
 export { SloStatusBadge } from './slo_status_badge';
 export { SloActiveAlertsBadge } from './slo_active_alerts_badge';
 export { SloStateBadge } from './slo_state_badge';
+export { SloValueBadge } from './slo_value_badge';
