+++ x-pack/solutions/observability/plugins/slo/public/pages/slo_details/components/overview/apm_indicator_overview.tsx
@@ -16,7 +16,7 @@ import {
 import React from 'react';
 import { useKibana } from '../../../../hooks/use_kibana';
 import { convertSliApmParamsToApmAppDeeplinkUrl } from '../../../../utils/slo/convert_sli_apm_params_to_apm_app_deeplink_url';
-import { OverviewItem } from './overview_item';
+import { DefinitionItem } from './definition_item';
 
 interface Props {
   slo: SLOWithSummaryResponse;
@@ -45,7 +45,7 @@ export function ApmIndicatorOverview({ slo }: Props) {
   } = indicator;
 
   return (
-    <OverviewItem
+    <DefinitionItem
       title={i18n.translate('xpack.slo.sloDetails.overview.apmSource', {
         defaultMessage: 'APM source',
       })}
