+++ x-pack/solutions/observability/plugins/slo/public/pages/slo_details/components/overview/overview.stories.tsx
@@ -10,11 +10,11 @@ import { ComponentStory } from '@storybook/react';
 
 import { KibanaReactStorybookDecorator } from '../../../../utils/kibana_react.storybook_decorator';
 import { buildSlo } from '../../../../data/slo/slo';
-import { Overview as Component, Props } from './overview';
+import { Definition as Component, Props } from './definition';
 
 export default {
   component: Component,
-  title: 'app/SLO/DetailsPage/Overview',
+  title: 'app/SLO/DetailsPage/Definition',
   decorators: [KibanaReactStorybookDecorator],
 };
