+++ x-pack/solutions/observability/plugins/slo/public/pages/slo_details/components/overview/synthetics_indicator_overview.tsx
@@ -14,7 +14,7 @@ import {
   syntheticsMonitorLocationQueryLocatorID,
 } from '@kbn/observability-plugin/common';
 import { useKibana } from '../../../../hooks/use_kibana';
-import { OverviewItem } from './overview_item';
+import { DefinitionItem } from './definition_item';
 
 interface Props {
   slo: SLOWithSummaryResponse;
@@ -41,14 +41,14 @@ export function SyntheticsIndicatorOverview({ slo }: Props) {
 
   const onMonitorClick = () => monitorLocator?.navigate({ configId: monitorId, locationId });
   const onLocationClick = () => regionLocator?.navigate({ locationId: location });
-  const showOverviewItem = name || location;
+  const showDefinitionItem = name || location;
 
-  if (!showOverviewItem) {
+  if (!showDefinitionItem) {
     return null;
   }
 
   return (
-    <OverviewItem
+    <DefinitionItem
       title={MONITOR_LABEL}
       subtitle={
         <EuiFlexGroup direction="row" alignItems="flexStart" gutterSize="s" responsive={false} wrap>
