+++ x-pack/solutions/observability/plugins/slo/public/pages/slo_details/slo_details.test.tsx
@@ -230,7 +230,6 @@ describe('SLO Details Page', () => {
     render(<SloDetailsPage />);
 
     expect(screen.queryByTestId('sloDetailsPage')).toBeTruthy();
-    expect(screen.queryByTestId('overview')).toBeTruthy();
     expect(screen.queryByTestId('sliChartPanel')).toBeTruthy();
     expect(screen.queryByTestId('errorBudgetChartPanel')).toBeTruthy();
     expect(screen.queryAllByTestId('wideChartLoading').length).toBe(2);
@@ -245,7 +244,6 @@ describe('SLO Details Page', () => {
     render(<SloDetailsPage />);
 
     expect(screen.queryByTestId('sloDetailsPage')).toBeTruthy();
-    expect(screen.queryByTestId('overview')).toBeTruthy();
     expect(screen.queryByTestId('sliChartPanel')).toBeTruthy();
     expect(screen.queryByTestId('errorBudgetChartPanel')).toBeTruthy();
     expect(screen.queryByTestId('errorRateChart')).toBeTruthy();
